Cache information responses from REST sources (#5726)
## Change
Store and use the `/information` response from REST sources per their
`cache-control` header.

At the base of this change is a new type of setting stream: `Encrypted`.
These settings are encrypted to the user and machine that the data is
stored on.

Based on the log file from running with and without a cached value, the
time between close log lines dropped from 189 ms to 8 ms for the msstore
source. This time saving should be in effect for the large majority of
uses of WinGet.