Shared build props (#5749)
## Change
Due to an issue with the internal build, it was realized that we could
be violating ODR due to a mismatch on the build configurations. This
change moves to using the MSBuild `Directory.Build.props` file to have
all projects use the same settings. It also adds a header using `#pragma
detect_mismatch` to attempt to detect the issue at compilation. This
header is then included by the errors header, which is used by at least
one TU in every project, if not all of them.