Add Font source group policy support (#5646)
<!-- To check a checkbox place an "x" between the brackets. e.g: [x] -->

This is for issue
[5644](https://github.com/microsoft/winget-cli/issues/5644)
* #5644

The purpose of this is to create a group policy definition in
anticipation of the upcoming font support so IT administrators can be
prepared for its inclusion and lead time.

It is not the intent of this checkin to actually implement the font
source, but to create the policy so enterprises can be better prepared
for it and have more lead time to adjust.

Changes:
* Added the Font Source group policy and associated tests and strings.

Verified build and groupPolicy tests.

- [x] I have signed the [Contributor License
Agreement](https://cla.opensource.microsoft.com/microsoft/winget-pkgs).
- [x] I have updated the [Release Notes](../doc/ReleaseNotes.md).
- [x] This pull request is related to an issue.

-----

###### Microsoft Reviewers: [Open in
CodeFlow](https://microsoft.github.io/open-pr/?codeflow=https://github.com/microsoft/winget-cli/pull/5646)