Support associating export units with packages in subdirectories of install location (#5859)
## Change
The primary motivation is to support directories below the install
location to contain configuration units that we will associate with the
package. This is achieved by refactoring the association logic from a
Package x Unit loop into a tree structure that is colored by package
install locations. This also has the benefit of making a O(N^2)
algorithm into an O(N).

Units are first inserted into the tree based on their file path. Then
the install location of each package is recorded onto that tree as well.
Finally, during the export of each package, all resources at the install
location and any that are descended from it but not under another
package are included.