Update platform toolset (#5882)
The solution does not build in VS 2026 because it defaults to using the
v140 toolset (VS 2015), which is no longer supported.
This PR adds the configuration to use the latest toolset v145 in VS 2026
(v18), and removes the options for v140 (VS 2015) and v141 (VS 2017).
I'm also setting this in the `Directory.Build.props` and removing the
configuration from the individual projects.
###### Microsoft Reviewers: [Open in
CodeFlow](https://microsoft.github.io/open-pr/?codeflow=https://github.com/microsoft/winget-cli/pull/5882)