Extract event log for potential crash info (#5807)
## Change
We are seeing occasional errors during test setup that might be a crash.
These started happening when 1.11.510 became the latest version. To
better understand this and any future crashes that might occur, grab the
Application event log from the pipeline.

Also grabs the GA winget logs since that is what is being used during
this potential crash.

Also enables the unit tests to write a minidump if they crash, since
that happened randomly during one of the builds for this PR.