Additional logging limitations and control (#5888)
## Change
Expands the limitations on log files and allows them to be user
configurable.

The existing limit was hardcoded to remove any files over 7 days old.
This was enforced by a background thread spun up during process start.
This same thread enforces the new limits on the aggregate logging
directory, while additional code in the file logger enforces the
individual file size limitation.

When new log lines would exceed the maximum file size for an individual
file, the stream is repositioned after the initialization portion of the
log, a log line is output to indicate the wrap, and the logs continue to
overwrite the older data.