Use hybrid CRT linkage instead of full static (#5913)
See https://aka.ms/hybridcrt for additional details.

When building ReleaseStatic, use the hybrid linkage model. This produces
a binary that is closer in size to the dynamic linkage than the full
static, relying on the UCRT binary that ships in Windows to provide the
core CRT functionality. As an example, the WindowsPackageManager.dll is
~90 KB larger as hybrid than dynamic, but is another ~160 KB larger as
static than hybrid.