Allow suppressing configuration output on test (#5794)
This PR adds the `--suppress-initial-details` argument to the `winget
configure test` subcommand. In addition, it makes it so the output can
be suppressed independent of accepting the agreements. Since the
agreement states that the user is responsible for knowing the
configuration they are applying, if they've decided to suppress those
details, they should be able to without having to also pass the switch
to accept the agreement. I've validated the agreement still appears