Enable Explicit toggling for sources (i.e. Enable/Disable) (#5904)
<!-- To check a checkbox place an "x" between the brackets. e.g: [x] -->

- [x] I have signed the [Contributor License
Agreement](https://cla.opensource.microsoft.com/microsoft/winget-pkgs).
- [x] I have updated the [Release Notes](../doc/ReleaseNotes.md).
- [x] This pull request is related to an issue.

-----
Closes #5797
This issue was related to enabling/disabling a source without removing
it. Functionally in WinGet this is already captured with the "Explicit"
property, where a source is not included unless it is explicitly
referenced. This is effectively equivalent to a "disabled" source. In
examining this problem we determined the best way to resolve this is to
allow editing of a source's properties, specifically the "Explicit"
property, without having to remove and re-add a source, which is
problematic with default sources. This PR implements that solution to
allow editing of a Source's properties, starting with the Explicit
property. This works on User defined sources and Default sources. In
terms of policy, Edit follows Remove; if you can remove a source you can
edit that source. If you cannot remove the source, then you cannot edit
the source either.

This PR Introduces a new experimental feature, "**sourceEdit**" which
has this new editing behavior. The user experience is not final so we
are keeping it behind sourceEdit for now for testing and refinement.

**New functionality: source edit subcommand**
This allows a source to be edited without removing/readding. This
currently only supports the "Explicit" property of a source, and
includes User sources and Default sources (such as the winget-font
source).

Aliases for this command: `set` and `config`

Example1:  `winget source edit -n winget --explicit true` 
This will make the winget source explicit.

Example2:  `winget source edit -n winget-font -e false`
This will make the winget-font source no longer explicit.

This interface is probably going to change later, which is one of the
reasons why we are using the experimental feature.

Key implementation detail is the addition of an "IsOverride" property to
the SourceDetailsInternal which functions similar to IsTombstone, only
instead of deleting the source it overrides some of the source's
properties, currently the "Explicit" property. This is needed for the
Default sources. For user sources the update is straightforward.

<img width="504" height="110" alt="image"
src="https://github.com/user-attachments/assets/d4d16dd2-61d7-460b-ba3e-310fbdb499e9"
/>

**Tested**
* Added CLI tests for the default source override validation
* Added E2E tests for the command validation
* Manual verification of the command functionality.


###### Microsoft Reviewers: [Open in
CodeFlow](https://microsoft.github.io/open-pr/?codeflow=https://github.com/microsoft/winget-cli/pull/5904)